requirements.txt
# this is the requirements.txt file...

requests

python-dotenv

pandas

plotly